CDP-716: Upload to Vimeo when token present
Uncommented /auth routes so we can use /auth/vimeo.
Added env var for enabling/disabling /auth/register: ALLOW_REGISTER=false/true
Added vimeo auth routes (/auth/vimeo and /auth/vimeo/callback).
Added getUnit function to abstractModel.
Added vimeo service for making various requests against the vimeo API using the Vimeo SDK.
Added site property to the stream object (to specify vimeo or cloudflare).
Added logic and functionality to the transfer ctrl that uploads to vimeo instead of cloudflare when an access token is on the header.
Also added delete functionality when a vimeo token is on the header and a post is deleted.