CDP-1173: Modify Video Route to Reflect Videos from Publisher
This is a preliminary commit for the changes happening with the introduction of publisher for videos.
Socket initialization and data flow control added.
Unnecessary video processing removed from transfer controller (which is still used by other routes).
Mediainfo replaced by ffprobe for obtaining size and duration of local and remote videos.
Validate schema split into two so that it can also be used by socket.